% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.rotate.R
\name{pre.rotate}
\alias{pre.rotate}
\title{Pre Rotate}
\usage{
pre.rotate(tree1, tree2)
}
\arguments{
\item{tree1}{First tree as phylo object. Will be represented in the left side tanglegram (Tree 1).}

\item{tree2}{Second tree as phylo object. Will be represented in the right side of tanglegram (Tree 2).}
}
\value{
List containing two rotated tree.
}
\description{
A function that reorder tree1 and tree2 so that their tips are in matching order.
The input and output tree will be phylo object
}
\examples{

tree1 <- read.tree('tree1.nwk') # Load tree 1
tree2 <- read.tree('tree2.nwk') # Load tree 2

plot(tree1)
plot(tree2)

treelist <- pre.rotate(tree1, tree2)

plot(treelist[[1]])
plot(treelist[[2]])

}
