% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.tanglegram.R
\name{triple.tanglegram}
\alias{triple.tanglegram}
\title{Triple Tanglegram}
\usage{
triple.tanglegram(
  tree1,
  tree2,
  tree3,
  column,
  sampletypecolors = NA,
  t2_pad = 0.5,
  t3_pad = 0.5,
  t2_y_scale = 1,
  t2_y_pos = 0,
  t3_y_scale = 1,
  t3_y_pos = 0,
  lab_pad = 0.05
)
}
\arguments{
\item{tree1}{First tree as ggtree object. Will be represented in the left side tanglegram (Tree 1).}

\item{tree2}{Second tree as ggtree object. Will be represented in the middle of tanglegram (Tree 2).}

\item{tree3}{Third tree as ggtree object. Will be represented in the right side of tanglegram (Tree 3).}

\item{column}{The column from meta data.frame associated with both trees which will be used to connect the tips.}

\item{sampletypecolors}{Named vector where names should correspond to the sample types, and the values are their associated colors.}

\item{t2_pad}{Tree 2 padding. Change this to adjust position of Tree 2. Default 0.5.}

\item{t3_pad}{Tree 3 padding. Change this to adjust position of Tree 3. Default 0.5.}

\item{t2_y_scale}{If Tree 2 is different size than Tree 1, then use this to adjust Tree 2 scale.}

\item{t2_y_pos}{If Tree 2 is different size than Tree 1, then use this to adjust their relative positions.}

\item{t3_y_scale}{If Tree 3 is different size than Tree 2, then use this to adjust Tree 3 scale.}

\item{t3_y_pos}{If Tree 3 is different size than Tree 2, then use this to adjust their relative positions.}

\item{lab_pad}{Add space after/before the tip-labels. It makes equidistant changes to the line x-positions. Default 0.05.}
}
\value{
ggplot object
}
\description{
A function that automatically draw tanglegram between shared tips in three phylogenies.
It can color the connecting lines by category in a column in the meta.
}
