% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.tanglegram.R
\name{common.tanglegram}
\alias{common.tanglegram}
\title{Common Tangleggram}
\usage{
common.tanglegram(
  tree1,
  tree2,
  column,
  tip_column,
  sampletypecolors = NA,
  t2_pad = 0.5,
  t2_y_scale = 1,
  t2_y_pos = 0,
  lab_pad = 0.05,
  text_width_factor = NULL,
  tiplab = FALSE,
  t2_tiplab_size = 3,
  t2_tiplab_pad = 0,
  link_colors = NULL,
  tip_colors = NULL
)
}
\arguments{
\item{tree1}{First tree as ggtree object. Will be represented in the left side tanglegram (Tree 1).}

\item{tree2}{Second tree as ggtree object. Will be represented in the right side of tanglegram (Tree 2).}

\item{column}{The column from meta data.frame associated with both trees which will be used to connect the tips.}

\item{tip_column}{Optional. The column from the meta data.frame used to color the tip points of the trees. Defaults to `column`.}

\item{sampletypecolors}{Backward-compatible alias for \code{link_colors}. A named
vector where names correspond to the categories in \code{column} and values are
colors. Existing calls using this argument continue to work.}

\item{link_colors}{Optional named vector used to manually color connecting
lines. Names must correspond to categories in \code{column}.}

\item{tip_colors}{Optional named vector used to manually color tip points.
Names must correspond to categories in \code{tip_column}.}

\item{t2_pad}{Tree 2 padding. Change this to adjust position of Tree 2. Default 0.5.}

\item{t2_y_scale}{If Tree 2 is different size than Tree 1, then use this to adjust Tree 2 scale.}

\item{t2_y_pos}{If Tree 2 is different size than Tree 1, then use this to adjust their relative positions.}

\item{lab_pad}{Add space after/before the tip-labels. It makes equidistant changes to the line x-positions. Default 0.05.}

\item{tiplab}{Boolean. Shows tip-labels of Tree 1. Default False. For showing tip-labels of Tree 1, add geom_tiplab() during defining the tree.}

\item{t2_tiplab_size}{Update Tree 2 tip label font-size. Default 3.}

\item{t2_tiplab_pad}{Add spacing between Tree 2 tip label and Tree 2 tip ending. Default 0.}
}
\value{
None
}
\description{
A function that automatically draw tanglegram between shared tips in two phylogeny.
It can color the connecting lines by category in a column in the meta.
For that, a named vector in R has to be provided. The names should correspond to the sample types, and the values are their associated colors.
Otherwise it will generate random colors from Viridis scale.
The inputs are ggtree object.
}
\examples{

meta=read.csv('tree_meta.csv', header=T) # Load meta
t1 <- read.tree('tree1.nwk') # Load tree 1

# Load tree 1 and use ggtree to annotate features
tree1 <- ggtree(t1) \%<+\% meta

# Load tree 2
t2 <- read.tree("tree2.nwk")
tree2 <- ggtree(t2) \%<+\% meta

# Make named vectors for connecting-line and tip-point colors
link_colors <- c("hospital" = "#4E79A7", "terrestrial" = "#F28E2B",
                 "animal" = "#E15759", "soil" = "#76B7B2")
tip_colors <- c("species_a" = "#59A14F", "species_b" = "#EDC948")
common.tanglegram(tree1, tree2, column_of_interest,
                  tip_column = species, link_colors = link_colors,
                  tip_colors = tip_colors, t2_pad = 1, tiplab = TRUE)


}
