% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.tanglegram.R
\name{common.tanglegram}
\alias{common.tanglegram}
\title{Common Tangleggram}
\usage{
common.tanglegram(
  tree1,
  tree2,
  column,
  sampletypecolors = NA,
  t2_pad = 0.3,
  t2_y_scale = 1,
  t2_y_pos = 0,
  lab_pad = 2,
  tiplab = FALSE,
  t2_tiplab_size = 1,
  t2_tiplab_pad = 0
)
}
\arguments{
\item{tree1}{First tree as ggtree object. Will be represented in the left side tanglegram (Tree 1).}

\item{tree2}{Second tree as ggtree object. Will be represented in the right side of tanglegram (Tree 2).}

\item{column}{The column from meta data.frame associated with both trees which will be used to connect the tips.}

\item{sampletypecolors}{Named vector where names should correspond to the sample types, and the values are their associated colors.}

\item{t2_pad}{Tree 2 padding. Change this to adjust position of Tree 2. Default 0.3.}

\item{t2_y_scale}{If Tree 2 is different size than Tree 1, then use this to adjust Tree 2 scale.}

\item{t2_y_pos}{If Tree 2 is different size than Tree 1, then use this to adjust their relative positions.}

\item{lab_pad}{Add space after/before the tip-labels. It makes equidistant changes to the line x-positions. Default 2.}

\item{tiplab}{Boolean. Shows tip-labels of Tree 1. Default False. For showing tip-labels of Tree 1, add geom_tiplab() during defining the tree.}

\item{t2_tiplab_size}{Update Tree 2 tip label font-size. Default 1.}

\item{t2_tiplab_pad}{Add spacing between Tree 2 tip label and Tree 2 tip ending.}
}
\value{
None
}
\description{
A function that automatically draw tanglegram between shared tips in two phylogeny.
It can color the connecting lines by category in a column in the meta.
For that, a named vector in R has to be provided. The names should correspond to the sample types, and the values are their associated colors.
Otherwise it will generate random colors from Viridis scale.
The inputs are ggtree object.
}
\examples{

meta=read.csv('tree_meta.csv', header=T) # Load meta
t1 <- read.tree('tree1.nwk') # Load tree 1

# Load tree 1 and use ggtree to annotate features
tree1 <- ggtree(t1)   \%<+\% meta +
  geom_tippoint(aes(color=species))
# Load tree 2
t2 <- read.tree("tree2.nwk")
tree2 <- ggtree(t2) \%<+\% meta

# Make a named vector
sampletypecolors <- sampletypecolors <- c("hospital" = "#4E79A7", "terrestrial" = "#F28E2B", "animal" = "#E15759", "soil" = "#76B7B2")
common.tanglegram(tree1, tree2, column_of_interest, sampletypecolors, t2_pad=1, tiplab = T)


}
